requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
lxml==6.0.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dotenv==1.1.1
python-pptx==1.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uuid==1.30
uvicorn==0.35.0
xlsxwriter==3.2.5
